requirements.txt

Flask==3.1.2
mysql-connector-python==9.5.0
requests==2.31.0
python-dotenv==1.0.1
Flask==3.1.2